In G_SaveMapState, fix 1024-byte write beyond array bound on 64-bit platforms.
This was caused by mismatched for types mapstate_t's animateptr[] and the global
one (int32_t vs. int32_t *).

Bump BYTEVERSION.

git-svn-id: https://svn.eduke32.com/eduke32@4125 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/sector.h
@@ -40,7 +40,7 @@ typedef struct {
     // this needs to have a copy of everything related to the map/actor state
     // see savegame.c
     int32_t animategoal[MAXANIMATES], animatevel[MAXANIMATES], g_animateCount;
-    int32_t animateptr[MAXANIMATES];
+    intptr_t animateptr[MAXANIMATES];
     int32_t lockclock;
     int32_t msx[2048], msy[2048];
     int32_t randomseed, g_globalRandom;
